Fix a typo in schema, add "failures" field to LogEntry, split monitor code into files
