+++ data/schema.go
@@ -347,4 +347,5 @@ type LogEntry struct {
 	Addr        string    `reform:"addr"`
 	Data        string    `reform:"data"`
 	Topics      LogTopics `reform:"topics"`
+	Failures    uint64    `reform:"failures"`
 }
